Rename files, fix pylint issues
