+++ flight/waypoint/move_to_test.py
@@ -3,24 +3,28 @@ Main driver code for moving drone to each waypoint
 """
 
 import asyncio
-import goto
-from mavsdk import System
 import logging
-from typing import List
 import sys
+
+from typing import List
+
+from mavsdk import System
+
+from flight.waypoint import goto
 
 
 async def run() -> None:
     """
     This function is a driver to test the goto function and runs through the
     given waypoints in the lats and longs lists at the altitude of 100.
-    Makes the drone move to each location in the lats and longs arrays at the altitude of 100 and
+    Makes the drone move to each location in the lats and longs arrays
+    at the altitude of 100 and
 
     Notes
     -----
-    Currently has 3 values in each the Lats and Longs array and code is looped and will stay in that loop
-    until the drone has reached each of locations specified by the latitude and longitude and
-    continues to run until forced disconnect
+    Currently has 3 values in each the Lats and Longs array and code is looped
+    and will stay in that loop until the drone has reached each of locations
+    specified by the latitude and longitude and continues to run until forced disconnect
     """
     # Put all latitudes, longitudes and altitudes into seperate arrays
     lats: List[float] = [37.9008502, 37.9008129, 37.8964543]
@@ -71,11 +75,9 @@ async def run() -> None:
         await asyncio.sleep(1)
 
 
+# Runs through the code until it has looped through each element of
+# the Lats and Longs array and the drone has arrived at each of them
 if __name__ == "__main__":
-    """
-    Runs through the code until it has looped through each element of
-    the Lats and Longs array and the drone has arrived at each of them
-    """
     try:
         loop = asyncio.get_event_loop()
         loop.run_until_complete(run())
